Move all the validation into a separate meta-module
