+++ src/proc/mod.rs
@@ -1,17 +1,14 @@
 //! Module processing functionality.
 
-pub mod analyzer;
 mod layouter;
 mod namer;
 mod terminator;
 mod typifier;
-mod validator;
 
 pub use layouter::{Alignment, Layouter};
 pub use namer::{EntryPointIndex, NameKey, Namer};
 pub use terminator::ensure_block_returns;
 pub use typifier::{ResolveContext, ResolveError, Typifier, TypifyError};
-pub use validator::{TypeFlags, ValidationError, Validator};
 
 impl From<super::StorageFormat> for super::ScalarKind {
     fn from(format: super::StorageFormat) -> Self {
